Added MainApp
Login window:
   if the login fails, an erro message will show
   else, the login window will close and the main app window will open.
The main app window:  2 tabs and a box on the left. In the first tab is a combobox from where to choose the age group(3 groups to choose from), and a table showing the results based on the choice made. Second tab is an advanced searh, also contains a table. The box on hte left lets you add a new registration, imput for first name, last name, age, and checkboxes for events 50/100, 100/1000, 1000/1500, also based on the age choosed. 2 buttons: the add button(which add the registration) and log out ( which close the main app window)